Restructure commands, update mux, improve error handling
